refactor: jpa 자동 테이블 생성 옵션을 create drop 대신 none으로 변경하고 직접 ddl 생성하도록 변경

+++ src/main/java/roomescape/application/dto/response/payment/PaymentResponse.java
@@ -3,7 +3,7 @@ package roomescape.application.dto.response.payment;
 import roomescape.domain.payment.Payment;
 
 public record PaymentResponse(
-        Long totalAmount,
+        int totalAmount,
         String paymentKey,
         String orderId,
         String status,
